refactor: PageController 내 메서드 역할별로 분리
