+++ src/main/java/wooteco/subway/web/PageController.java
@@ -54,34 +54,4 @@ public class PageController {
     public String servicePage() {
         return "service/index";
     }
-
-    @GetMapping(value = "/join", produces = MediaType.TEXT_HTML_VALUE)
-    public String joinPage() {
-        return "service/join";
-    }
-
-    @GetMapping(value = "/login", produces = MediaType.TEXT_HTML_VALUE)
-    public String loginPage() {
-        return "service/login";
-    }
-
-    @GetMapping(value = "/logout", produces = MediaType.TEXT_HTML_VALUE)
-    public String logoutPage() {
-        return "service/logout";
-    }
-
-    @GetMapping(value = "/mypage", produces = MediaType.TEXT_HTML_VALUE)
-    public String myPage() {
-        return "service/mypage";
-    }
-
-    @GetMapping(value = "/mypage-edit", produces = MediaType.TEXT_HTML_VALUE)
-    public String myPageEdit() {
-        return "service/mypage-edit";
-    }
-
-    @GetMapping(value = "/favorites", produces = MediaType.TEXT_HTML_VALUE)
-    public String favoritesPage() {
-        return "service/favorite";
-    }
 }
